# src/main/java/io/swagger/api/AccountApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.4.13).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package io.swagger.api;

import io.swagger.model.LoginRequest;
import io.swagger.model.LoginRet;
import io.swagger.annotations.*;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
@javax.annotation.Generated(value = "io.swagger.codegen.languages.SpringCodegen", date = "2020-04-25T02:02:49.831Z")

@Api(value = "account", description = "the account API")
@RequestMapping(value = "")
public interface AccountApi {

    @ApiOperation(value = "用户名密码校验", nickname = "login", notes = "用户名密码校验接口", response = LoginRet.class, tags={ "account", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "OK", response = LoginRet.class),
        @ApiResponse(code = 400, message = "Invalid RequestBody supplied"),
        @ApiResponse(code = 404, message = "RequestBody not found") })
    @RequestMapping(value = "/account/login",
        produces = { "application/json" }, 
        consumes = { "application/json" },
        method = RequestMethod.POST)
    default ResponseEntity<LoginRet> login(@ApiParam(value = "用户登陆login信息" ,required=true )  @Valid @RequestBody LoginRequest body) {

        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
    }

}