# src/main/java/io/swagger/api/OtherApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.4.13).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package io.swagger.api;

import io.swagger.model.ContactListGetRet;
import io.swagger.model.ContactListSetRet;
import io.swagger.model.ContactSetRequest;
import io.swagger.annotations.*;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
@javax.annotation.Generated(value = "io.swagger.codegen.languages.SpringCodegen", date = "2020-04-25T02:02:49.831Z")

@Api(value = "other", description = "the other API")
@RequestMapping(value = "")
public interface OtherApi {

    @ApiOperation(value = "社交媒体联系列表", nickname = "contactlistget", notes = "社交媒体联系列表", response = ContactListGetRet.class, tags={ "other", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "OK", response = ContactListGetRet.class) })
    @RequestMapping(value = "/other/get_contactlist",
        produces = { "application/json" }, 
        consumes = { "application/json" },
        method = RequestMethod.GET)
    default ResponseEntity<ContactListGetRet> contactlistget() {

        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
    }


    @ApiOperation(value = "设置社交媒体联系", nickname = "contactlistset", notes = "设置社交媒体联系接口", response = ContactListSetRet.class, tags={ "other", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "OK", response = ContactListSetRet.class) })
    @RequestMapping(value = "/other/set_contactlist",
        produces = { "application/json" }, 
        consumes = { "application/json" },
        method = RequestMethod.POST)
    default ResponseEntity<ContactListSetRet> contactlistset(@ApiParam(value = "" ,required=true )  @Valid @RequestBody ContactSetRequest body) {

        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
    }

}