# src/main/java/io/swagger/api/PageApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.4.13).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package io.swagger.api;

import io.swagger.model.PageNameGetRet;
import io.swagger.model.PageNameSetRequest;
import io.swagger.model.PageNameSetRet;
import io.swagger.annotations.*;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
@javax.annotation.Generated(value = "io.swagger.codegen.languages.SpringCodegen", date = "2020-04-25T02:02:49.831Z")

@Api(value = "page", description = "the page API")
@RequestMapping(value = "")
public interface PageApi {

    @ApiOperation(value = "获取个人主页名称", nickname = "pagenameget", notes = "获取个人主页名称接口", response = PageNameGetRet.class, tags={ "page", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "OK", response = PageNameGetRet.class),
        @ApiResponse(code = 400, message = "Invalid RequestBody supplied"),
        @ApiResponse(code = 404, message = "RequestBody not found") })
    @RequestMapping(value = "/page/get_name",
        produces = { "application/json" }, 
        consumes = { "application/json" },
        method = RequestMethod.GET)
    ResponseEntity<PageNameGetRet> pagenameget();


    @ApiOperation(value = "设置个人主页名称", nickname = "pagenameset", notes = "设置个人主页名称接口", response = PageNameSetRet.class, tags={ "page", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "OK", response = PageNameSetRet.class) })
    @RequestMapping(value = "/page/set_name",
        produces = { "application/json" }, 
        consumes = { "application/json" },
        method = RequestMethod.POST)
    ResponseEntity<PageNameSetRet> pagenameset(@ApiParam(value = "个人主页名称" ,required=true )  @Valid @RequestBody PageNameSetRequest body);

}